Add ansible_ssh_private_key_file to the SSH test
We're waiting for the OpenStack VMs to be able to SSH into before
proceeding with the DNS playbook (because until 2.3 Ansible doesn't
handle that gracefully).

But if the private key for the SSH connection was only specified using
the `--private-key` argument to `ansible-playbook`, the test was
failing because it didn't take that possibility into the account.

Now it does.

Fixes #355