Refactor disk creation to instance templates
There is currently a bug in GCE API that when you specify a non-boot disk sourceImage as a family it will instead use the sourceImage from the boot disk.  To workaround this we don’t use a family to specify this sourceImage even though it is more appropriate to do so.