Add roles to create and delete empty image (workaround)
GCE API does not allow you to specify empty disks in instance templates.  This is a workaround to that limitation.  The version of cloudilb currently available as an RPM on my build system also doesn’t allow me to specify a family for this image.  The impact of this is limited because GCE API has a bug where as if we specify the image using the family it doesn’t work as expected.