GCP Ref. Arch.: add the possibility to attach to an existing VPC
The reference architecture scripts for GCP are currently taking care of creating
a dedicated VPC and subnet for hosting the created OpenShift cluster.

There are some use cases where one may want to have the scripts attach the VM
to a pre-existing VPC and subnet:
— create an OpenShift cluster on a shared VPC defined in a distinct GCE project;
— create an OpenShift cluster on VPC/subnet previously connected via VPN to a
  remote network.

This change adds the possibility to make the GCP reference architecture scripts
attach the instances to a pre-existing VPC/subnet rather than creating a new one.