Allow installation of OCP 3.9 on GCP
As OCP 3.9 is already released this commit updates the major version
from 7 to 9 and related variables and paths from the
config.yaml.example file.

Some review and modifications to other reference architectures'
playbooks/roles might be needed.

This particular commit contains changes needed only to get it working
on GCP.

- increase openshift_major_version to 9
- fix path for the openshift-ansible installation playbook
- use openshift-ansible prerequisites playbook
- add openshift_facts playbook to gce ref arch inst playbook
- fix library path to point to lib_utils role in openshift-ansible
- fix metrics & logging related installation variables
- add GCP cloud provider multizone configuration
- add GCP cloud provider vars (prefix & project)
- change openshift.common.is_atomic to openshift_is_atomic conditions